Add comment to fieldset 2 and fill out fieldset 3
I just added a comment to the second fieldset, to make me remember to
fill it out. i have to learn about dropbox, thats why this field stays
a little empty for now.

The third fieldset is a radio button, that I have already learned about.
Here i filled out with a yes-no question, gave an id-for, name, value and
type attribute to it, let only select just one option.